some fixes

+++ components/cinna-slack/chat/components/process.js
@@ -182,31 +182,54 @@ var imageSearch = function(data,token,callback){
           // if(res.responses && res.responses[0].logoAnnotations && res.responses[0].logoAnnotations[0]){
           //   searchTerms.push(res.responses[0].logoAnnotations[0].description);
           // }
-   
-          //text detection
-          if(res.responses && res.responses[0].textAnnotations && res.responses[0].textAnnotations[0] && res.responses[0].textAnnotations[0].locale == 'en' || res.responses[0].textAnnotations[0].locale == 'es' || res.responses[0].textAnnotations[0].locale == 'fr'){ //only processing english, spanish, french right now
-            var textEx = res.responses[0].textAnnotations[0].description;
-            textEx = textEx.replace(/(\r\n|\n|\r)/gm," "); //remove line breaks
-            textEx = textEx.replace(/[\u0250-\ue007]/g, ''); //remove non-latin characters
-            textEx = textEx.replace(/^(.{30}[^\s]*).*/, "$1"); //limit # of words
-            searchTerms.push(textEx);
-          }
+
+          // //text detection
+          if(res.responses && res.responses[0].textAnnotations && res.responses[0].textAnnotations[0]){ //only processing english, spanish, french right now
+            if (res.responses[0].textAnnotations[0].locale == 'en' || res.responses[0].textAnnotations[0].locale == 'es' || res.responses[0].textAnnotations[0].locale == 'fr'){
+              var textEx = res.responses[0].textAnnotations[0].description;
+              textEx = textEx.replace(/(\r\n|\n|\r)/gm," "); //remove line breaks
+              textEx = textEx.replace(/[\u0250-\ue007]/g, ''); //remove non-latin characters
+              textEx = textEx.replace(/^(.{30}[^\s]*).*/, "$1"); //limit # of words
+              searchTerms.push(textEx);          
+            }
+          } 
 
           //label detection
-          else if(res.responses && res.responses[0].labelAnnotations && res.responses[0].labelAnnotations[0]){ 
-            searchTerms.push(res.responses[0].labelAnnotations[0].description);
+          if (searchTerms.length < 1){
+
+            if(res.responses && res.responses[0].labelAnnotations && res.responses[0].labelAnnotations[0]){
+              searchTerms.push(res.responses[0].labelAnnotations[0].description);
+            }
+
             //lol this code is awful
-            if(res.responses[0].labelAnnotations[1]){
+            if(res.responses && res.responses[0].labelAnnotations && res.responses[0].labelAnnotations[1]){
               searchTerms.push(res.responses[0].labelAnnotations[1].description);
             }
+
             //lol this code is awful
-            if(res.responses[0].labelAnnotations[2]){
+            if(res.responses && res.responses[0].labelAnnotations && res.responses[0].labelAnnotations[2]){
               searchTerms.push(res.responses[0].labelAnnotations[2].description);
             }
 
+            //lol this code is awful
+            if(res.responses && res.responses[0].labelAnnotations && res.responses[0].labelAnnotations[3]){
+              searchTerms.push(res.responses[0].labelAnnotations[3].description);
+            }
+
+            //lol this code is awful
+            if(res.responses && res.responses[0].labelAnnotations && res.responses[0].labelAnnotations[4]){
+              searchTerms.push(res.responses[0].labelAnnotations[4].description);
+            }            
           }
 
-          callback(Array.from(new Set(searchTerms)).join(" ")); //remove dupes and make into string on return
+          // check for search terms
+          if(searchTerms.length > 0){
+            console.log(searchTerms);
+            callback(Array.from(new Set(searchTerms)).join(" ")); //remove dupes and make into string on return
+          }
+          else {
+            callback();
+          }
 
           fs.unlinkSync(savePath); //remove temp image
